refactor: split main into focused modules
Extract core logic from main.rs into dedicated modules (claude, cli,
error, flags, git, prompt, spinner) for better separation of concerns.
Add thiserror crate for proper error handling instead of String-based
errors. This improves code organization and maintainability.
