+++ Cargo.toml
@@ -18,6 +18,7 @@ clap = { version = "4", features = ["derive"] }
 indicatif = "0.17"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+thiserror = "2"
 
 [profile.release]
 lto = "fat"
